+++ src/DirectDebitGateway.php
@@ -7,20 +7,18 @@
  * Company: Pronamic
  *
  * @author Reüel van der Steege
- * @version 1.0.0
+ * @version 1.0.1
  * @since 1.0.0
  */
 class Pronamic_WP_Pay_Extensions_Give_DirectDebitGateway extends Pronamic_WP_Pay_Extensions_Give_Gateway {
 	/**
-	 * Constructs and initialize an Direct Debit gateway
+	 * Constructs and initialize Direct Debit gateway.
 	 */
 	public function __construct() {
-		$this->id = 'pronamic_pay_direct_debit';
-
-		$this->name = __( 'Direct Debit', 'pronamic_ideal' );
-
-		$this->payment_method = Pronamic_WP_Pay_PaymentMethods::DIRECT_DEBIT;
-
-		parent::__construct();
+		parent::__construct(
+			'pronamic_pay_direct_debit',
+			__( 'Direct Debit', 'pronamic_ideal' ),
+			Pronamic_WP_Pay_PaymentMethods::DIRECT_DEBIT
+		);
 	}
 }