+++ CHANGELOG.md
@@ -7,7 +7,13 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 -
 
+## [1.0.1] - 2016-04-12
+- Improved error handling.
+- Refactored class construct gateways.
+- Empty unused get URL functions.
+
 ## 1.0.0 - 2016-03-23
 - First release.
 
-[unreleased]: https://github.com/wp-pay-extensions/give/compare/1.0.0...HEAD
+[unreleased]: https://github.com/wp-pay-extensions/give/compare/1.0.1...HEAD
+[1.0.1]: https://github.com/wp-pay-extensions/give/compare/1.0.0...1.0.1